+++ squarelet/organizations/querysets.py
@@ -47,7 +47,7 @@ class OrganizationQuerySet(models.QuerySet):
 
 class MembershipQuerySet(models.QuerySet):
     def get_viewable(self, user):
-        """Returns memberships in either public organizations or any org the user is a member of"""
+        """Returns memberships in public orgs or any org the user is a member of"""
         return self.filter(Q(organization__private=False) | Q(organization__users=user))
