+++ squarelet/users/viewsets.py
@@ -19,7 +19,6 @@ from rest_framework.response import Response
 
 # Squarelet
 from squarelet.core.mail import send_mail
-from squarelet.core.permissions import DjangoObjectPermissionsOrAnonReadOnly
 from squarelet.oidc.permissions import ScopePermission
 from squarelet.organizations.models import Membership, Plan
 from squarelet.users.models import User
@@ -27,7 +26,6 @@ from squarelet.users.serializers import (
     PressPassUserSerializer,
     UserReadSerializer,
     UserWriteSerializer,
-    PressPassUserMembershipsSerializer,
 )
 
 
@@ -110,23 +108,6 @@ class PressPassUserViewSet(
             return self.request.user
         else:
             return super().get_object()
-
-
-class PressPassUserMembershipViewSet(
-    mixins.ListModelMixin,
-    viewsets.GenericViewSet,
-):
-    queryset = Membership.objects.none()
-    serializer_class = PressPassUserMembershipsSerializer
-    permission_classes = (DjangoObjectPermissions,)
-    lookup_field = "user__uuid"
-
-    def get_queryset(self):
-        user = get_object_or_404(
-            User,
-            uuid=self.kwargs["user_uuid"],
-        )
-        return user.memberships.get_viewable(self.request.user)
 
 
 class PressPassRegisterView(RegisterView):
